@@ -1,11 +1,14 @@
 import 'dart:convert';
 
-abstract class ServiceWebRequestInterface {
-  Future get(String url, {Map<String, String>? headers});
+import 'service_web_response_interface.dart';
 
-  Future post(String url,
+abstract class ServiceWebRequestInterface {
+  Future<ServiceWebResponseInterface> get(String url,
+      {Map<String, String>? headers});
+
+  Future<ServiceWebResponseInterface> post(String url,
       {Object? body, Map<String, String>? headers, Encoding? encoding});
 
-  Future delete(String url,
+  Future<ServiceWebResponseInterface> delete(String url,
       {Object? body, Map<String, String>? headers, Encoding? encoding});
 }
